Updated get_exp type

+++ src/map/src/combat/combat.cpp
@@ -286,8 +286,8 @@ void Combat::update(EntitySystem& entitySystem, Entity entity, uint32_t dt) {
           // End exp out algo
 
           auto exp_required = get_exp_to_level(attackerLevel.level+1);
-          while (attackerLevel.xp >= exp_required) {
-            int32_t new_exp = attackerLevel.xp - exp_required;
+          while ((int64_t)attackerLevel.xp >= exp_required) {
+            int64_t new_exp = attackerLevel.xp - exp_required;
             attackerLevel.xp = new_exp;
             ++attackerLevel.level;
             exp_required = get_exp_to_level(attackerLevel.level+1);
@@ -504,7 +504,7 @@ void Combat::revive(EntitySystem& entitySystem, Entity entity, const RoseCommon:
   entitySystem.teleport_entity(entity, std::get<1>(dest), std::get<2>(dest), std::get<0>(dest));
 }
 
-int32_t Combat::get_exp_to_level(int level) {
+int64_t Combat::get_exp_to_level(int level) {
   if ( level > 210 )
     level = 210;
